+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.2] - 2022-02-08
+
+Drop support for postgres replay queue. Batch insert operations in
+postgres backend.
+
 ## [3.0.0] - 2022-02-06
 
 Optimized top collection for systems with millions of processes.